Move pages folder to components folder.
